requirements.txt
# LangGraph 和 LangChain 核心
langgraph>=0.2.0,<0.3.0
langchain>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0  # Excel 读取
xlrd>=2.0.0      # 老版本 Excel 支持

# 数据验证
pydantic>=2.0.0

# 持久化存储
aiosqlite>=0.19.0

# 环境变量管理
python-dotenv>=1.0.0

# 代码执行隔离（可选，用于安全执行生成的代码）
# RestrictedPython>=6.0.0
